\name{mapuloh-package}
\alias{mapuloh-package}
\alias{mapuloh}
\docType{package}
\title{
\packageTitle{mapuloh}
}
\description{
\packageDescription{mapuloh}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mapuloh}
\packageIndices{mapuloh}}
\author{
\packageAuthor{mapuloh}

Maintainer: \packageMaintainer{mapuloh}
}
\references{}
\keyword{ package }
\seealso{}
\examples{}
